pyproject.toml
[tool.poetry]
name = "fixms"
version = "0.2.9"
description = ""
authors = ["Alec Thomson (S&A, Kensington WA) <alec.thomson@csiro.au>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.9"
numpy = "<2"
astropy = ">=5"
python-casacore = "*"
tqdm = "*"
importlib-resources = {version="*", python = "<3.9"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = ">=5.12.0"
black = ">=23.3.0"
sphinx = "*"
numpydoc = "*"
myst_parser = "*"
autoapi = "*"
sphinx-autoapi = "*"
pytest = "*"
pre-commit = "*"

[tool.poetry.extras]
dev = [
    "ruff",
    "sphinx",
    "numpydoc",
    "myst_parser",
    "autoapi",
    "sphinx-autoapi",
    "pydata-sphinx-theme",
    "pytest",
    "pre-commit",
]


[tool.poetry.scripts]
fix_ms_corrs = "fixms.fix_ms_corrs:cli"
fix_ms_dir = "fixms.fix_ms_dir:cli"
fix_ms = "fixms.fix_ms:cli"
